backend/requirements.txt
fastapi==0.115.0
motor==3.3.2
pymongo>=4.5,<4.8
uvicorn==0.30.0
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.0.8
celery==5.4.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0
python-multipart==0.0.9
slowapi==0.1.9
httpx==0.27.2
email-validator>=2.0
google-genai>=1.0
